dont delete cts if is benchmark test

+++ installation/scripts/testing.sh
@@ -166,8 +166,10 @@ kubectl get cts ${suiteName} -oyaml
 echo "Generate JUnit test summary (${JUNIT_REPORT_PATH})"
 kyma test status "${suiteName}" -ojunit | sed 's/ (executions: [0-9]*)"/"/g' > "${JUNIT_REPORT_PATH}"
 
-
-kubectl delete cts ${suiteName}
+if [[ ! "${BENCHMARK}" == "true" ]]
+then
+  kubectl delete cts ${suiteName}
+fi
 
 printImagesWithLatestTag
 latestTagExitCode=$?
